+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `getPeripheral` xplat API
+
 ### Changed
 
 - Scanning now is ability with infinite scanning possibilities